♻️ Refactor: modal client 이동
