+++ src/app/workspace/[workspaceId]/sns-event-assistant/@snsEventAssistantGeneralModal/(.)creating-event/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import CreatingEventModalClient from './CreatingEventModalClient';
+import CreatingEventModalClient from '../../../../../../features/sns-event-assistant/components/modal-client/CreatingEventModalClient/CreatingEventModalClient';
 
 const CreatingEventModalPage = () => {
   return (
