+++ src/app/workspace/[workspaceId]/sns-event-assistant/@snsEventAssistantGeneralModal/(.)new-event/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import NewEventModalClient from './NewEventModalClient';
+import NewEventModalClient from '../../../../../../features/sns-event-assistant/components/modal-client/CreatingEventModalClient/NewEventModalClient';
 
 const NewEventModalPage = () => {
   return (
