+++ src/features/sns-event-assistant/components/SnsLinkItem/SnsLinkItem.client.tsx
