+++ src/features/sns-event-assistant/components/SnsLinkItem/SnsLinkItem.type.ts
